[BASE] Fix implementation of TraitDescriptor::for_each.
It is necessary to traverse all traits instead of find first alternative which contains a trait that is set.
Let us consider an example, where some group is TraitAlternative<TraitUnion<T1, T2>, TraitUnion<T1, T3>> and T1, T3 is set.
The first alternative it TraitUnion<T1, T2> it contains T1 which is set, but we could not stop evaluation of this group because T1 is also presented in second alternative.