[BCL, CMake] Do not store build directory in the package registry.
`export(PACKAGE <name>)` changes global state.
I think that such behaviour is not evident for BCL users.
It seems that if BCL is used without installation then it should be local.
So, CMake package registry should not be updated.
Furthermore, this command uses path to the build tree.
So, this command is not necessary for installation.